requirements.txt
flake8
bandit
coverage
pytest
pytest-cov
pylint
biopython
plotly
